requirements.txt
pymupdf>=1.18.0,<1.23.0
pillow>=8.0.0,<10.0.0
numpy>=1.21.0,<1.25.0
openpyxl>=3.0.0,<4.0.0
pyinstaller>=4.0,<6.0
